+++ Demo/MainWindow.xaml.cs
@@ -19,6 +19,7 @@ public partial class MainWindow
         [
             new ControlItem("ChapterButton", new ChapterButtonControl()),
             new ControlItem("ChapterCheckBox", new ChapterCheckBoxControl()),
+            new ControlItem("ChapterRepeatButton", new ChapterRepeatButtonControl()),
             new ControlItem("ChapterSplitButton", new ChapterSplitButtonControl()),
             new ControlItem("ChapterToggleButton", new ChapterToggleButtonControl()),
             new ControlItem("ChapterToggleSwitch", new ChapterToggleSwitchControl())